requirements.txt
openai 
pyyaml